Fix HUD scaling issues
Instead of using fixed-size HUDs that don't fit on all
screens, stretch the HUD images to fit the screen.
Assume that all screens use a close enough aspect
ratio, and that minor distortion is tolerable.

Add margins around the title card so that it fits
full-screen without distortion.

Also prescale the ghost text overlay texture to match
the size of the text content textures, so that we don't
rely on MT's built-in scaling and use the full resolution
of both textures.