Add drop shadows to titles to improve readability
Currently using computed shadows to make it easier
to modify the source images, though this requires
hard-coding the image sizes into the code, making it
impossible for texture packages to change the size.
In the future a separate texture build process should
be considered to create images with pre-baked drop
shadows given unshadowed source images.