Voice acted ghost dialog
Each line spoken by the ghost now has an associated
sound that plays for it.

Sounds are played positionally, but with a spatial effect
to give the ghostly voice an otherworldly quality.

Background music is reduced in volume globally (for
now) to accomodate the dialog.  In the future we may
want to "duck" the volume down below the dialog
instead.

Also, the time the text is displayed is still hard-coded
and fixed, and not matched to the dialog timing.