Stateless music management
Instead of having to signal to the ambiance/music
system whenever a transition between states is
needed (which is error prone, as time traveling while
near a ghost in the source/target time shows), have
the music system simply scan for ghosts that should
affect the soundtrack.

If an "endgame" ghost is present at all, play the
ending area song.  Otherwise, ghosts must be
near the player to affect the music, and otherwise
with no ghosts, play the normal theme.

Rather than starting/stopping music all the time,
it's easier to manage just having persistent handles
for all songs and fade them in/out while keeping
the loops running; this also gives us the "resume
song from current position" feature we wanted for
passing by a ghost quickly (Fix #23).

Ducking music for speech is controlled by a
field on the ghost entity now, so only applies when
the ghost is within range.

This system is also inherently multiplayer compat,
in the event we want to make the game cooperative.