Force disable damage, fix sky/physics changes
Even if enable_damage is disabled in the UI, it can still
be forcibly enabled in settings or command line, so make
the player immortal and hide the health-related HUDs.

The sky was only being set to black on the player's first
join; make it use the specified all-black sky on every
session even after reloading the world.