Rudimentary item uniqueness check
Each of the 15 collectable items should be "unique",
but many, such as the treasures, are loaded in with
schematics in each era, even if the player already
took one from any era.  Since many items cannot be
lost or discarded (treasures, stones) we can tell if the
player ever took it based on it being present in
inventory, and remove it from the world in that case
(immediately upon time travel, and as a cleanup ABM
for old saved games and other leaks).  For the plants
we can't guarantee the seed won't exist in multiple
places in multiple eras, but we can minimize the risk
the player can actually collect a duplicate by triggering
the additional copies to self-destruct if any one copy
is collected.