Place items in their assigned slots
When picking up a node, overwrite the correct item slot that the
item is assigned to.  This effectively mitigates any duplication
issues because duplicate nodes in the world will just overwrite
the existing copy of the item in that slot.

Also make the initial items (book, letter) honor their item slots
and move control of this into the item definitions.